labs/lab08-langgraph-persistence/requirements.txt
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
langgraph>=0.0.30
langgraph-checkpoint>=0.0.1
python-dotenv>=1.0.0
typing-extensions>=4.0.0

# Database drivers
aiosqlite>=0.19.0
asyncpg>=0.29.0
azure-cosmos>=4.5.1
pyodbc>=5.0.1
sqlalchemy>=2.0.23
